LeetCode/src/test/java/org/redquark/onlinejudges/leetcode/list/RemoveLinkedListElementsTest.java
package org.redquark.onlinejudges.leetcode.list;

import org.junit.jupiter.api.Test;
import org.redquark.onlinejudges.leetcode.common.ListNode;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNull;

public class RemoveLinkedListElementsTest {

    private final RemoveLinkedListElements testObject = new RemoveLinkedListElements();

    @Test
    public void testRemoveElements() {
        ListNode head = new ListNode(1);
        head.next = new ListNode(2);
        head.next.next = new ListNode(6);
        head.next.next.next = new ListNode(3);
        head.next.next.next.next = new ListNode(4);
        head.next.next.next.next.next = new ListNode(5);
        head.next.next.next.next.next.next = new ListNode(6);
        int val = 6;
        ListNode expected = testObject.removeElements(head, val);
        assertEquals(1, expected.val);
        assertEquals(2, expected.next.val);
        assertEquals(3, expected.next.next.val);
        assertEquals(4, expected.next.next.next.val);
        assertEquals(5, expected.next.next.next.next.val);

        head = null;
        val = 1;
        assertNull(testObject.removeElements(head, val));

        head = new ListNode(7);
        head.next = new ListNode(7);
        head.next.next = new ListNode(7);
        head.next.next.next = new ListNode(7);
        val = 7;
        assertNull(testObject.removeElements(head, val));
    }
}